Topic model import update
Topic model was hardcoded with the date in it.

Updated to now look for topic_model_current so we do not have to commit every time a new model is trained.

There is functionality that would allow for "modular" imports, where we could look and import the most recent model, but I do no think that is worth the effort.